requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
asyncio==3.4.3
attrs==18.2.0
chardet==3.0.4
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
python-dotenv==0.10.1
telepot==12.7
typing-extensions==3.7.2
urllib3==1.24.1
yarl==1.3.0
